+++ src/main/java/com/github/camilormz/finalreality/model/weapon/types/Staff.java
@@ -2,8 +2,11 @@ package com.github.camilormz.finalreality.model.weapon.types;
 
 import com.github.camilormz.finalreality.model.weapon.AbstractWeapon;
 import com.github.camilormz.finalreality.model.weapon.IMagicalWeapon;
+import com.github.camilormz.finalreality.model.weapon.IWeapon;
 import com.github.camilormz.finalreality.model.weapon.WeaponType;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.Objects;
 
 /**
  * A class that holds all the information of a staff.
@@ -25,5 +28,26 @@ public class Staff extends AbstractWeapon implements IMagicalWeapon {
     @Override
     public int getMagicDamage() {
         return this.magicDamage;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof IMagicalWeapon)) {
+            return false;
+        }
+        final IMagicalWeapon weapon = (IMagicalWeapon) o;
+        return this.getName().equals(weapon.getName()) &&
+               this.getDamage() == weapon.getDamage() &&
+               this.getWeight() == weapon.getWeight() &&
+               this.getMagicDamage() == weapon.getMagicDamage();
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(this.getName(), this.getDamage(), this.getWeight(), this.getType(),
+                            this.getMagicDamage());
     }
 }